Visual-Haptic
This is one of my project for my Master study. I'm creating an
integrated setup for visual-haptic virtual environment, which is using a
Haptic device with Virtual Reality. The user will experience the
interaction with a soft object using a commercial force – feedback
device (impedance control) and a head mounted display.

The Haptic device used in this project is the Novint Falcon. By using
Chai3D to initialize the device in the programming code, I integrate the
Haptic device function with Virtual Reality by using Oculus Rift.